+++ views/components/tables/top-visitors.php
@@ -73,7 +73,7 @@ use WP_Statistics\Utils\Url;
                                         ]);
                                         ?>
 
-                                        <?php $campaign = Url::getParam('?' . $page['query'], 'utm_campaign'); ?>
+                                        <?php $campaign = sanitize_text_field((string) Url::getParam('?' . $page['query'], 'utm_campaign')); ?>
                                         <?php if ($campaign) : ?>
                                             <span class="wps-campaign-label wps-tooltip" title="<?php echo esc_attr__('Campaign:', 'wp-statistics') . ' ' . esc_attr($campaign); ?>"><?php echo esc_html($campaign); ?></span>
                                         <?php endif; ?>